requirements.txt
alembic==1.0.11
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-MySQL==1.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.14
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
mysqlclient==1.4.2.post1
pkg-resources==0.0.0
protobuf==3.9.0
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.6
Werkzeug==0.15.5
